Display wake notification when staying in a thread
Notifications are not displayed when you receive a new message within a
chat. If the user would put the device to sleep while viewing a chat
then the any notifications received from that chat would be still be
suppressed but the device screen would still wake.

Now when the device is asleep the notification is always shown.